requirements.txt
PyJWT[cryptography]==2.4.0
